add the test to test_dataloader2.py
